requirements.txt
torch
torchvision
Pillow
opencv-python
matplotlib
segmentation-models-pytorch
numpy
